Foundation Fix
Fixed some issues with the components not scaling appropriately.
Added the HOC to each of the components so they are created with a motion section around each for future animations
Added scaling starts which grows dynamically with the content within a component.
Made a sample test for components with loads of resources to make sure everything lines up appropriately.